+++ libopenage/curve/demo/gui.h
@@ -1,4 +1,4 @@
-// Copyright 2015-2017 the openage authors. See copying.md for legal info.
+// Copyright 2017-2017 the openage authors. See copying.md for legal info.
 
 #pragma once
