+++ libopenage/curve/events/events.cpp
