--- libopenage/curve/events/events2.h
@@ -1,34 +1,0 @@
-// Copyright 2017-2017 the openage authors. See copying.md for legal info.
-
-#pragma once
-
-#include "../curve.h"
-
-#include "../../util/hash.h"
-
-#include <list>
-#include <memory>
-#include <queue>
-#include <unordered_set>
-#include <map>
-
-// Forward declarations
-namespace openage {
-class State;
-namespace curvepong {
-int demo();
-}
-}
-
-namespace openage {
-namespace curve {
-
-class Event;
-class EventQueue;
-class EventClass;
-class EventManager;
-
-
-
-
-}} // openage::curve
