+++ libopenage/curve/test/CMakeLists.txt
@@ -3,6 +3,6 @@ link_libraries(tube)
 add_sources(libopenage
 	test_container.cpp
 	test_curve_types.cpp
-	test_events2.cpp
+	test_events.cpp
 	test_serialization.cpp
 )
